[BED-109] Extract slow iron configuration to independent class
Took 8 minutes
